Split out multi class file
