+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/EnumValuesExtension/Animal.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Toolkit.Uwp.SampleApp.Enums
+{
+    public enum Animal
+    {
+        Cat,
+        Dog,
+        Bunny,
+        Parrot,
+        Squirrel
+    }
+}
